fix clearPlaceholder
clearPlaceholder frequently caused errors, and those errors were being suppressed.

I realized that it was only necessary to clear the placeholder when text was inserted into it, so I made a correction to check before calling it.

This should prevent the errors from being suppressed.

This correction is a bit redundant, but I plan to make a more efficient correction later.