bugfix: Fixed an issue where images were not displayed correctly when changing layouts
When updateLayout is executed, a new page is created internally, and the previous object information is deleted. Therefore, it was necessary to delete the preload information as well.

However, since no new uploads were made for images similar to those previously included in the layout, I also made changes to copy images originating from Markdown within uploadLayout to prevent issues.